+++ backend/internal/repository/gpu_repository.go
@@ -2,10 +2,26 @@ package repository
 
 import (
 	"context"
+	"time"
 
 	"github.com/kamil7430/gpu-share/backend/internal/model"
 )
 
 type GpuRepository interface {
 	GetDeviceStatusById(ctx context.Context, id string) (*model.DeviceStatus, error)
+}
+
+type mockGpuRepository struct{}
+
+func NewMockGpuRepository() GpuRepository {
+	return &mockGpuRepository{}
+}
+
+func (m *mockGpuRepository) GetDeviceStatusById(ctx context.Context, id string) (*model.DeviceStatus, error) {
+	return &model.DeviceStatus{
+		TemperatureC:       69,
+		UtilizationPercent: 69,
+		MemoryUsedMb:       6969,
+		LastHeartbeat:      time.Date(2005, 4, 2, 21, 37, 0, 0, time.UTC),
+	}, nil
 }
