--- backend/internal/repository/mock_gpu_repository.go
@@ -1,23 +1,0 @@
-package repository
-
-import (
-	"context"
-	"time"
-
-	"github.com/kamil7430/gpu-share/backend/internal/model"
-)
-
-type MockGpuRepository struct{}
-
-func NewMockGpuRepository() GpuRepository {
-	return &MockGpuRepository{}
-}
-
-func (m *MockGpuRepository) GetDeviceStatusById(ctx context.Context, id string) (*model.DeviceStatus, error) {
-	return &model.DeviceStatus{
-		TemperatureC:       69,
-		UtilizationPercent: 69,
-		MemoryUsedMb:       6969,
-		LastHeartbeat:      time.Date(2005, 4, 2, 21, 37, 0, 0, time.UTC),
-	}, nil
-}
